#1: Bug fix - Backpropagation assigns same value to both players
Backpropagation should assign "1" as value to a node in case of win
and "0" in case of loss. The current implementation assigns after a
rollout the same value to all nodes. It should however assign different
values, depending on who is the currentPlayer. Right now for one player
not the best moves are picked during the Selection phase, but the worst
moves. Because of this the algorithm picks rather random moves.